% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateIQR.R
\name{CalculateIQR}
\alias{CalculateIQR}
\title{CalculateIQR}
\usage{
CalculateIQR(data)
}
\arguments{
\item{data}{a vector of your data}
}
\description{
This function calculates the distance of each data point from the nearest quartile boundary, expressed in units of the Interquartile Range (IQR). It returns a vector where each element indicates how many IQR units a data point deviates from either the first quartile (Q1) or third quartile (Q3), depending on its position relative to the interquartile range.
}
