% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIC_ODGC.R
\name{PIC_ODGC}
\alias{PIC_ODGC}
\title{PICODGC}
\usage{
PIC_ODGC(my_df, X, Y, my_tree, IQR_threshold = 1.5)
}
\arguments{
\item{my_df}{dataframe of traits}

\item{X}{trait1 name}

\item{Y}{trait2 name}

\item{my_tree}{phylogenetic tree}

\item{IQR_threshold}{IQR threshold of outlier}
}
\value{
PICODGC model
}
\description{
Perform PICODGC
}
